requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
cached-property==2.0.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
dnspython==2.7.0
et_xmlfile==2.0.0
exchangelib==5.5.1
frozenlist==1.7.0
idna==3.10
isodate==0.7.2
lxml==6.0.0
multidict==6.6.3
oauthlib==3.3.1
openpyxl==3.1.5
propcache==0.3.2
pycparser==2.22
Pygments==2.19.2
pyspnego==0.11.2
python-dotenv==1.1.1
python-owasp-zap-v2.4==0.1.0
requests==2.32.4
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
yarl==1.20.1
zaproxy==0.4.0
